Dynamic Link Progress
Discovered an implemented one possible start to a mostly physics
(non-Kinematic) solution to tether/link mechanic:

Still experimenting with flaws/advantages and seeing what can be learned
from this implementation. It may seem like a Kinematic solution is the
most controlled way to go about this. Right now hold the down arrow key
to keep the update function running. It is set like this for testing
purposes, however it shows how a possible "stiff" tether might be
implemented in the future.

*Features in development/debugging:*
- 2 player joined, and able to inflict forces onto each other
- Link Dynamically changes size, effectively only limiting the max
distance.
- Link will be able to be switched between kinematic modes, so
developers may experiment which is best.